requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
numpy
yandex_cloud_ml_sdk
langchain
